Refactor metric layer map component
